+++ python/changelog.md
@@ -11,6 +11,9 @@
 ### Enhancements
 
 * Added check for chunk length 0 sent from the server to avoid full agent crash on running.
+* Using requests session
+* Added option for using proxies
+* Added HTTP Basic Auth support
 
 ## 0.1.8 -> 0.2.0